PR: rename package localdaemon to locald
